Added a status check for login
The 'login' link on the nav bar will now change depending on the status of the user. If they are logged in, it will display 'logged in' and the href changes to blank (could change to account). If they are not logged in, it links to the login page and displays "login/signup".

Also updated the Navbar html to give it JS functionality